feat: implement multi-step Create Account form with separate components for each step
